+++ src/components/SortableItem.tsx
@@ -1,0 +1,29 @@
+import React from "react";
+import {useSortable} from "@dnd-kit/sortable";
+import {CSS} from "@dnd-kit/utilities";
+
+interface SortableItemProps {
+    id: string
+    children: React.ReactNode
+}
+export const SortableItem: React.FC<SortableItemProps> = ({ id, children }) => {
+    const { attributes, listeners, setNodeRef, transform, transition } =
+        useSortable({ id })
+
+    const style = {
+        transform: CSS.Transform.toString(transform),
+        transition,
+    }
+
+    return (
+        <div
+            ref={setNodeRef}
+            style={style}
+            {...attributes}
+            {...listeners}
+            className="task"
+        >
+            {children}
+        </div>
+    )
+}
