+++ src/index.tsx
@@ -4,8 +4,8 @@ import { BlockEntity } from '@logseq/libs/dist/LSPlugin.user'
 import { createRoot } from 'react-dom/client'
 import { renderToString } from 'react-dom/server'
 
-import { Kanban } from './components/Kanban'
-import KanbanDnd from './components/KanbanDnd'
+import { Kanban } from './features/Kanban'
+import KanbanDnd from './features/KanbanDnd'
 import { createNormalBoard } from './helpers/create-normal-board'
 import { createNormalBoardWithQuery } from './helpers/create-normal-query-board'
 import { createQueryBoard } from './helpers/create-query-board'
